+++ core/platform/aws/src/main/kotlin/org/http4k/aws/AwsCredentials.kt
@@ -1,3 +1,6 @@
 package org.http4k.aws
 
-data class AwsCredentials(val accessKey: String, val secretKey: String, val sessionToken: String? = null)
+data class AwsCredentials(val accessKey: String, val secretKey: String, val sessionToken: String? = null) {
+    override fun toString() =
+        "AwsCredentials(accessKey=$accessKey, secretKey=****, sessionToken=${sessionToken?.let { "****" }})"
+}